Adding tokenization
